Changed compiler options to be more strict and pedantic

+++ pieces.h
@@ -22,7 +22,7 @@
 #include "types.h"
 
 
-#define NO_PIECE	(-1)
+#define NO_PIECE	((unsigned int)(-1))
 
 char get_piece_label(piece_id_t piece);
 colour_t get_colour(piece_id_t piece);
